Move Name and Tag classes to commons package
