+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -20,8 +20,8 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.name.Name;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.tag.Tag;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
 import seedu.address.model.vendor.Vendor;
