+++ src/main/java/seedu/address/logic/parser/EditCommandParser.java
@@ -16,7 +16,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.EditCommand.EditVendorDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.commons.tag.Tag;
 
 /**
  * Parses input arguments and creates a new EditCommand object
