+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.name.Name;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.tag.Tag;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
 import seedu.address.model.vendor.Vendor;
