+++ src/test/java/seedu/address/logic/parser/CreateCommandParserTest.java
@@ -20,9 +20,9 @@ import org.junit.jupiter.api.Test;
 import seedu.address.logic.commands.CreateCommand;
 import seedu.address.logic.commands.CreateEventCommand;
 import seedu.address.logic.commands.CreateVendorCommand;
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.event.Date;
 import seedu.address.model.event.Event;
-import seedu.address.model.name.Name;
 import seedu.address.model.vendor.Vendor;
 import seedu.address.testutil.VendorBuilder;
