+++ src/test/java/seedu/address/logic/parser/CreateVendorCommandParserTest.java
@@ -33,8 +33,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.CreateVendorCommand;
-import seedu.address.model.name.Name;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.tag.Tag;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
 import seedu.address.model.vendor.Vendor;
