+++ src/test/java/seedu/address/logic/parser/ParserUtilTest.java
@@ -14,9 +14,9 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.tag.Tag;
 import seedu.address.model.event.Date;
-import seedu.address.model.name.Name;
-import seedu.address.model.tag.Tag;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
