+++ src/test/java/seedu/address/model/ModelManagerTest.java
@@ -21,9 +21,9 @@ import javafx.collections.ObservableList;
 import javafx.collections.SetChangeListener;
 import javafx.util.Pair;
 import seedu.address.commons.core.GuiSettings;
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.event.Date;
 import seedu.address.model.event.Event;
-import seedu.address.model.name.Name;
 import seedu.address.model.vendor.NameContainsKeywordsPredicate;
 import seedu.address.model.vendor.Vendor;
 import seedu.address.testutil.AddressBookBuilder;
