+++ src/test/java/seedu/address/model/event/UniqueEventListTest.java
@@ -10,9 +10,9 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.event.exceptions.DuplicateEventException;
 import seedu.address.model.event.exceptions.EventNotFoundException;
-import seedu.address.model.name.Name;
 
 public class UniqueEventListTest {
     private final UniqueEventList uniqueEventList = new UniqueEventList();
