+++ src/test/java/seedu/address/storage/JsonAdaptedEventTest.java
@@ -9,8 +9,8 @@ import static seedu.address.testutil.TypicalEvents.FIONA;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.event.Date;
-import seedu.address.model.name.Name;
 
 public class JsonAdaptedEventTest {
     private static final String INVALID_NAME = "R@chel";
