+++ src/test/java/seedu/address/storage/JsonAdaptedVendorTest.java
@@ -12,7 +12,7 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.name.Name;
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
