+++ src/test/java/seedu/address/testutil/EventBuilder.java
@@ -1,8 +1,8 @@
 package seedu.address.testutil;
 
+import seedu.address.model.commons.name.Name;
 import seedu.address.model.event.Date;
 import seedu.address.model.event.Event;
-import seedu.address.model.name.Name;
 
 /**
  * A utility class to help with building Event objects.
