+++ src/test/java/seedu/address/testutil/VendorBuilder.java
@@ -3,8 +3,8 @@ package seedu.address.testutil;
 import java.util.HashSet;
 import java.util.Set;
 
-import seedu.address.model.name.Name;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.tag.Tag;
 import seedu.address.model.util.SampleDataUtil;
 import seedu.address.model.vendor.Description;
 import seedu.address.model.vendor.Phone;
